Move OccuranceConstraint to root namespace
